Remove dependency on hex_lit
We have a macro in `hex_conservative` that can parse hex strings, it
differs from `hex_lit` in that it uses heap allocated memory instead of
the stack.

Note also this PR uses the macro in non-test code but when we wrote it
in `hex-conservative` we called it `test_hex_unwrap`.

Remove the dependency on `hex_lit` and use the macro from
`hex-conservative`.