psbt: Remove usage of PrivateKey
High level reason for this patch: compressed keys are legacy, we should
make them second class citizens. There is no need for the `psbt` module
to be treating legacy keys as first class citizens.

The PSBT defines a serialization format, whether or not we use
serialize private keys using compressed format or not is an
implementation detail. We can use the keys for secp and remove all usage
of the uncompressed serialization.

Caveat: Is our PSBT serialization supposed to be readable by other
software? I highly doubt it is but just asking in case.

This breaks serde serialization of PSBTs also, so this makes the `psbt`
serialization format backwards incompatible.