Do not guarantee WASM with MSRV
We currently only test WASM using the stable toolchain. This means we do
not guarantee that WASM works with MSRV.

Currently the WASM build is breaking MSRV and there is such a knot of
dependencies that I cannot figure out how to pin them.

Explicitly do not guarantee WASM works with MSRV by adding a section to
the readme of `hashes` a well as the main workspace readme.

Remove the WASM related dev-dependencies from the `hashes` manifest and
add them dynamically during the CI run.