fuzz: Update generate-files
Recently we merged some changes to the manifest of the `fuzz` crate that
were created by dependabot, however we neglected to update the manifest
generator script we have.

Update `generate-files.sh` to include the github actions changes
currently in `fuzz/Cargo.toml`.

Do not run the `generate-files.sh` script, will be done as a separate
patch.