Add Witness::p2tr_keyspend function
Add a function for creating the witness when doing a key path spend for
a P2TR output.

This mirrors what we did for P2WPKH when adding `Witness::p2wpkh`.

We use `keyspend` as a single identifier because it appears to be a
commonly used word [0].

Includes update to the taproot signing example to use the new constructor.

Notes:

This patch highlights that the `taproot::Signature` serialization API
has two problems:

- It is non-uniform with the ECDSA one (which has `SerializedSignature`)
- It requires an allocation when we could serialize on the stack

[0] https://bitcoinsearch.xyz/?q=keyspend&size=n_50_n