Check in CI that commits conform to mkchlog format
To make changlog generation easier we want to put changes into commit
messages such that users are not bothered by internal details and
changelog can be generated automatically. A new tool called `mkchlog`
can solve this.

This change adds a github action that checks commits to conform to the
mkchlog format and configures appropriate start commit as well as basic
template.

changelog:
	section: dev
	inherit: title
	title-is-enough: true