Add new sha256t::HashEngine type
Currently we use a type alias for the `sha256t::HashEngine` type. This
means that users can inadvertently create a hash engine that has not had
the tag prehashed into it.

Add a new `sha256t::HashEngine` type that wraps a `sha256::HashEngine`.
Implement `Hash` by calling through to the inner `sha256::HashEngine`.

Include a constructor for creating a hash engine from a
`sha256::HashEngine` that has had the tag already hashed into it.