Impl push-based consensus decoding for a few types
This implements `bitcoin_consensus_encoding::Decode` for `Transaction`
and the types used inside it to demonstrate how `push_decode` can be
used to abstract over the byte source (sync/async, std/no-std). The
change is designed to be incremental so the exisitng code continues to
use old traits, we're just adding new capabilities that may later
replace the current encoding approach.