Add integer operations for locktime Height types
To improve ergonomics it would be good to support a subset of integer
operations that make sense for the two locktime heights.

- Add `checked_*` and `saturating_*` functions for add, sub, div, mul
- Add functions to increment (add 1) and decrement (sub 1) for each
 `Height` type. Call through to the respective `saturating_` function.
- Implement `ops::*` traits, `Add`, `Mul`, `Sub`, `Div` (as well as
  `AddAssign` and friends).

It makes sense to add a number of blocks onto both absolute and relative
heights so we implement, for example:

- `ops::Add<u32> for Height`
- `ops::Add<Height> for u32`
- `ops::AddAssign<u32> for Height`

It also makes sense to add a `relative::Height` to an `absolute::Height`
but not two absolute heights together. so we implement:

- relative height:
    - `ops::Add for Height`
    - `ops::AddAssign for Height`

- absolute height:
   - `ops::Add<relative::Height> for Height`
   - `ops::AddAssign<relative::Height> for Height`

Same for other operations.

We can implement `Add<u16>` and since the max value of
`relative::Height` and `u16` is the same we get the same unsurprising
behavior in relation to overflow and panic.

For the `absolute` type `absolute::Height::MAX` is less that the inner
ints max value.

As suggested here:

 https://github.com/rust-bitcoin/rust-bitcoin/issues/1639#issuecomment-1805770413

implement `ops` traits by unconditionally panicing if overflow occurs.

Done as part of #1639 and also to support the calculate next target stuff.