Remove network from PrivateKey and add WifKey
We would like to remove the network field from `PrivateKey` because the
network is to an attribute of a key, it is used solely to create and
parse the Wallet Import Format (WIF) string.

Add a new `WifKey` type. Remove the network field and WIF stuff from
`PrivateKey`.

Note, with this applied:

1. We now derive serde traits on the `PrivateKey` (includes update to
  serde regression test).
2. We do not implement `FromStr` and `Display` on the `PrivateKey`.

(2) seems reasonable, printing secret keys should not be a common thing
to do, so making it a bit hard seems like a good idea. If devs really
want to print secret keys they can get the bytes with
`wif.key.secret_bytes()` and print those in hex or whatever. `Debug`
calls through to secp which obfuscate the secrets.