Make encode module private
The `consensus::encode` module is really a code organisation thing and
should not be part of the pubic API. Also `encode::Error` is a strange
name for an error that happens when decoding, `consensus::Error` is
nicer on the eyes.

Make the `consensus::encode` module private and re-export everything in
it (explicitly).

This is a breaking change - obviously.