Remove non_exhaustive from Network
Recently we added the `non_exhaustive` attribute to `Network` in an
attempt to make the type forward compatible. This has proven to be an
annoyance to downstream users because there is no obvious recourse for
the unknown variant and matching on `Network` is common.

Note that if/when a new network is added by Core it is a big deal so we
can and should bump our version when we add support for it.

Remove `#[non_exhaustive]` from the `Network` type.

Fix: #2225