Rename `PublicKey` to `LegacyPublicKey`
We would like to shorten the name `CompressedPublicKey` into `PublicKey`
since legacy things are second-class in this library. However doing this
directly would make upgrading the library very painful, maybe even
error-prone. To aid with this we first just rename `PublicKey` and leave
a deprecated type alias in its place.