Introduce `script::Template`
When analyzing scripts the user had to call one of the `is_` methods and
then manually extract the correct subslice and possibly convert it into
another type. This was annoying as one had to look up the offsets,
handle witness length restrictions and other things. Further the code
doing this was far from the code implementing the `is_` methods making
audits harder.

To improve the situation we define two new enums: `script::Template` and
`script::RawTemplate`. The `Script::raw_template` method returns
`RawTemplate` which indicates which specific script template (if any) is
recognized and carriess the associated data with it. These associated
data are just raw bytes with correct length and, in case of SegWit,
correct version (for unknown versions). The `RawTemplate` can be
converted to `Template` which carries strongly typed, validated data
semantically identical to those in `RawTemplate`. Consumers are expected
to primarily use `Template` (and `Script::template` method) but
`RawTemplate` is still exposed for analysis or performance-sensitive
applications.

The change also makes use of this new feature in implementation of
`Address` which greatly simplifies the code. While the total lines
increase, hopefully this will be offset when more applications start
using `Template` instead of manual parsing.