Backport address conversion functions
Backport the address conversion functions from PR #1765. This is not a
stragiht cherry-pick because the `address` module has changed
substantially.

In order to be able to grab the two required conversion functions we
need to also create the `AddressInner` type so that we are guaranteed
the same layout when casting in unsafe code.

The `network` field is accessed once outside the `address` module so add
a getter method for it but use `.0` inside the `address` module.