CI: Move linting to top level script
Instead of running the linter in each individual CI script we can just
run it once from the top level script and exit. This has the benefit of
being quicker with no loss of linting coverage. Note this does remove
sanity builds from the linting job, this does not reduce overall
coverage of the whole CI workflow.