CI: Split out lint and docs jobs
The CI script is overly complex to reason about because of how it
interacts with the github actions. In an effort to make it more obvious
exactly what is going on and what is failing when CI is red split the
linting and docs stuff out into separate jobs. In the CI script once
these are run exit. This has the added advantage of making these jobs
fast to run and fast to fail, hopefully making the dev feedback loop
quicker.