CI: Move ASAN stuff to top level crate
The ASAN tests do not use `cargo --locked` so can be run from the top
level CI script.

This patch _increases_ coverage because the job that enables DO_ASAN was
only being run in `hashes/` - now we run it in `fuzz/` as well.

Note please I'm not very knowledgeable about ASAN stuff or what exactly
is going on in these commands but I remove `--lib` (and added `+nightly`
to be explicit) in order to get the it working in `fuzz/`.