CI: Introduce CARGO variable
The individual CI scripts are expected to be run after copying the
recent/minimal lock files into place. As such all `cargo` commands
should be run with `--locked` otherwise the lockfile will be updated. To
save us making mistakes in the future introduce a variable `CARGO` and
always use it instead of `cargo`.

This patch changes the current behaviour because in some places the
sanity checks are being run without using `--locked` - this is a bug.