Revert "Add effective_value method"
This reverts commit d69d62822d448c7ef33df85f441bbb8c5faec7b7.

This commit added a function to the public API that is not of a high
enough standard for a 1.0 release. It objectively makes the code base
worse and subjectively makes our job of getting to 1.0 harder.

The function can be implemented outside of `rust-bitcoin` so no user is
hobbled if we do not include this function.