Rename `WitnessProgram` to `SegwitScriptPubkey`
`WitnessProgram` was originally added with the incorrect name because it
was thought the inaccuracy is OK. However apart from slight confusion
this also prevents us from adding the actual witness program type which
could be useful in some situations. Thus we now rename it and deprecate
to give people the opportunity to update their crates, then we can add
`WitnessProgram` back with different semantics.