Enforce `$source: std::error::Error` in `write_err`
Sometimes people accidentally use `write_err` where `write` should be
used or forget to implement `std::error::Error` for their types. So far
we've relied mainly on review to catch these but it's difficult to see
them and as this commit demonstrates one case slipped in #1400.

This commit add a check to `write_err` that enforces implementation of
`std::error::Error + 'static` on types passed to `write_err` as
`$source`. This doesn't trigger on `no_std` builds but does in `std`
builds. While this would normally be weird in this case it's OK because
`std` is tested.

There are still cases where this check is a false positive so the
original macro is preserved as `write_err_unchecked_source`.

This change simultaneusly fixes the horrible bug introduced in #1400
which would cause infinite recursion when displaying error on `no_std`.