Improve taproot builder conversion errors
The two conversion functions on `TaprootBuilder` currently return the
same error type even though it contains a variant that cannot be
returned for `try_into_node_info`.

Split the error up in the customary fashion.

Please note this PR shows a new favouring of named fields in struct
errors instead of using a tuple struct. The struct has a single public
field so `non_exhaustive` is used.