Move height and time types to units
Move the bitcoin relative and absolute height and time types over to the
units crate.

This patch introduces a regression because it removes the ability to
parse height or time types from hex. This is because the `hex` crate is
being actively worked on to handle the prefix stuff that is currently in
`bitcoin::string` and which was used by the height and time types.